Refactor: Update navigation components and configuration
- Import `redirects` from the new `redirects.ts` file instead of `redirects.js` for consistency.
- Introduce a new `redirects.ts` file that exports an empty `redirects` array, preparing for future redirect configurations.
- Change the `href` type in `NavigationMainLink` from `BaseLinkProps['href']` to `LinkProps<PathValues>['href']` for better type safety with Next.js links.
- Replace the `SmartTextLink` component with the `Link` component from Next.js to streamline navigation and improve performance.
