+++ apps/filecoin-site/src/app/_components/Navigation/NavigationMainLink.tsx
@@ -1,16 +1,13 @@
 'use client'
 
-import { type LinkProps } from 'next/link'
+import Link, { type LinkProps } from 'next/link'
 import { usePathname } from 'next/navigation'
 
 import { clsx } from 'clsx'
 
-import type { BaseLinkProps } from '@filecoin-foundation/ui/BaseLink'
-import { SmartTextLink } from '@filecoin-foundation/ui/TextLink/SmartTextLink'
 import type { TouchTarget } from '@filecoin-foundation/utils/types/touchTargetType'
 
 import type { PathValues } from '@/constants/paths'
-import { BASE_DOMAIN } from '@/constants/siteMetadata'
 
 const TOUCH_TARGET: TouchTarget = {
   touchAreaPadding: 'px-5 py-6',
@@ -34,7 +31,7 @@ const mobileStyle = clsx(
 type NavigationLinkProps = {
   on: 'mobile' | 'desktop'
   label: string
-  href: BaseLinkProps['href']
+  href: LinkProps<PathValues>['href']
   onNavigate?: LinkProps<PathValues>['onNavigate']
 }
 
@@ -48,9 +45,8 @@ export function NavigationMainLink({
   const isActive = pathname.startsWith(href.toString())
 
   return (
-    <SmartTextLink
+    <Link
       href={href}
-      baseDomain={BASE_DOMAIN}
       aria-label={`Go to ${label} page`}
       aria-current={isActive}
       className={clsx(
@@ -61,6 +57,6 @@ export function NavigationMainLink({
       {...rest}
     >
       {label}
-    </SmartTextLink>
+    </Link>
   )
 }
